fix race condition for key expiry. remove unneeded EXISTS calls.
This commit makes two changes. The first is the use of a single set of logic to
increment and expire a key in Redis regardless of whether the key previously
exists or not. This avoids a race condition where a key is INCR without calling
EXPIRE in the case where the key is expired by Redis between checking for its
existence and calling only INCR on the key.

Separately, this commit moves from using an EXIST and a GET call to only making
a GET call where possible, since GET already returns an explicit "nil" error
when key does not exist.